refactor(journal): extract JournalReviewTimestamp (#299)
JournalReviewView inlined a ~16-line "Logged At" timestamp card plus
the static `timestampFormatter` and `formattedTimestamp` helpers
(~10 lines) — together they spread the timestamp concern across three
parts of the file. Lift everything into a focused
JournalReviewTimestamp view that owns its own DateFormatter and
exposes an optional `timestamp:` parameter (defaults to `Date()` so
the existing call-site behavior is preserved).

- JournalReviewView: 258 → 232 lines (-26).
- New Views/Journal/JournalReviewTimestamp.swift (~52 lines) with a
  #Preview macro and `.foregroundStyle` adopted.

Zero behavioral change: same medium-date + short-time formatting,
same uppercase caption label, same elevated-card-fill background.
